Updated the intended purpose
After chatting to Tom, established that the main point is not just automation, it is mainly to remove the burden from the clinician to have to think of calculating bleeding/ischaemia risk in the first place. This can also be achieved by just requiring the clinician to input info, and then flagging the HBR risk in auto generated patient notes.